Wip .NET Core upgrade
MdnMonitor compiling and passing tests.  Using Common.Logging.Framework, finally getting away from diagnostics.nlog static creation an that custom ioc.  It is a painful and slow process.  I don't like that MdnMonitor has a dependency on config.store but that is another day.  Kind of thinking of using Steeltoe so that is creeping in some.  Might use it's discovery and, ... maybe MdnMonitor can write to RabbitMQ or Kafka as a way to schedule timeout checks.  Not sure if Steeltoe is staying up with the times.  This is all because I am comparing this MdnMonitor to the one in the Java RI Micro-Service implementation using Spring Boot.

Other misc refactoring.  Eventually I can get the branch in shape breaking dependencies and moving from the old SmtpAgentConfig.Xml injection in favor of the standard ASP.NET DI and appsettings.json technique.  I have written some ugly code putting up with XML deserialization.