Wip 6.0 and EF Core 6.0
Brought in Postgresql and ran all tests accept for Mdn tests.  It looks like Postgresql supports with clause.  But SqlServer uses top where Postgresql uses Limit.  But the real miss match is Postgress doesn't seem to like the mixed case table and column names, it is forcing me to quote everywhere.  It has a default schema of public.  I now have a config.store.Npgsql and config.store.sqlserer where the entities and maps live.  They have the same namespace as the config.store projet.  I am using a custom configuration setting to conditionally compile in the correct mappings.  Then an environment variable for unit test to conditionally run the correct provider flavor.  Not sure how far I take this but it is interesting to see the differences between different database providers.  For Postgres I am using  docker-compose file to bring up Postgres and PgAdmin.  Super easy to be up and running fast.

Anyway needed to save my work.
And moved a bit further on the config.serer.rest project and tests.